+++ test/basic.js
@@ -12,7 +12,7 @@ var wskey = shed.generateKey();
 var CLIENT;
 var HTTP;
 var SERVER;
-
+var SERVER_ERR;
 
 function failOnMessage(msg) {
 	throw new Error('received unexpected message: ' + msg.toString());
@@ -29,7 +29,14 @@ test('setup server', function (t) {
 	HTTP = http.createServer();
 	HTTP.listen(9554);
 	HTTP.on('upgrade', function (req, socket, head) {
-		SERVER = shed.accept(req, socket, head);
+		try {
+			SERVER = shed.accept(req, socket, head, false,
+			    ['test1', 'test2']);
+		} catch (err) {
+			socket.end('HTTP/1.1 400 Bad Request\r\n' +
+			    'Connection: close\r\n\r\n');
+			return;
+		}
 		SERVER.on('text', failOnMessage);
 		SERVER.on('binary', failOnMessage);
 		SERVER.on('end', failOnEnd);
@@ -53,6 +60,8 @@ test('setup client', function (t) {
 	req.on('upgrade', function (res, socket, head) {
 		socket.setNoDelay(true);
 		CLIENT = shed.connect(res, socket, head, wskey);
+		t.strictEqual(CLIENT.getProtocol(), null);
+		t.strictEqual(SERVER.getProtocol(), null);
 		CLIENT.on('text', failOnMessage);
 		CLIENT.on('binary', failOnMessage);
 		CLIENT.on('end', failOnEnd);
@@ -137,16 +146,74 @@ test('server sends BINARY', function (t) {
 	SERVER.send(new Buffer('hello'));
 });
 
-test('teardown', function (t) {
+test('teardown connection', function (t) {
 	CLIENT.removeListener('end', failOnEnd);
 	SERVER.removeListener('end', failOnEnd);
 
 	CLIENT.on('end', function (code, reason) {
 		t.equal(code, 'NORMAL', 'normal close');
 		t.equal(reason, 'test ended', 'server sent reason');
-		HTTP.close();
 		t.end();
 	});
 
 	SERVER.end('test ended');
+
+	CLIENT = null;
+	SERVER = null;
+});
+
+test('negotiate supported subprotocol', function (t) {
+	var options = {
+		port: 9554,
+		hostname: '127.0.0.1',
+		headers: {
+			'connection': 'upgrade',
+			'upgrade': 'websocket',
+			'Sec-WebSocket-Key': wskey,
+			'Sec-WebSocket-Version': 13,
+			'Sec-WebSocket-Protocol': 'foobar, test1, test2'
+		}
+	};
+	var req = http.request(options);
+	req.end();
+	req.on('upgrade', function (res, socket, head) {
+		socket.setNoDelay(true);
+		var client = shed.connect(res, socket, head, wskey);
+		t.strictEqual(client.getProtocol(), 'test1');
+		t.strictEqual(SERVER.getProtocol(), 'test1');
+
+		SERVER.removeListener('end', failOnEnd);
+		SERVER.once('end', function () {
+			t.end();
+		});
+		client.end('done');
+	});
+});
+
+test('negotiation failure', function (t) {
+	var options = {
+		port: 9554,
+		hostname: '127.0.0.1',
+		headers: {
+			'connection': 'upgrade',
+			'upgrade': 'websocket',
+			'Sec-WebSocket-Key': wskey,
+			'Sec-WebSocket-Version': 13,
+			'Sec-WebSocket-Protocol': 'foobar, aaaa'
+		}
+	};
+	var req = http.request(options);
+	req.end();
+	req.on('upgrade', function (res, socket, head) {
+		t.fail();
+	});
+	req.on('response', function (res) {
+		t.equal(res.statusCode, 400);
+		t.end();
+	});
+});
+
+test('teardown server', function (t) {
+	HTTP.close();
+	t.end();
 });